update distance in start click

+++ dotNet5781_3169_8515/BL/BO/bus.cs
@@ -11,9 +11,9 @@ namespace BO
         public const int FULL_TANK = 1200;
 
         public TimeSpan time { get; set; }
-        public int fuel { get; set; }
-        public int distance { get; set; }
-        public int totalDistance { get; set; }
+        public double fuel { get; set; }
+        public double distance { get; set; }
+        public double totalDistance { get; set; }
         public bool dangerous { get; set; }
         public string plateNumber { get; set; }
         public DateTime registrationDate { get; set; }
